rename: 컴포넌트명 RegisterFirstStep, RegisterSecondStep, RegisterFirstStepForm으로 명확하게 수정
